Crate burn::backend::autodiff

Burn Autodiff

This autodiff library is a part of the Burn project. It is a standalone crate that can be used to perform automatic differentiation on tensors. It is designed to be used with the Burn Tensor crate, but it can be used with any tensor library that implements the Backend trait.

Modules

grads

Gradients module.

ops

Operation module.

Macros

testgen_ad_abs

testgen_ad_adaptive_avg_pool1d

testgen_ad_adaptive_avg_pool2d

testgen_ad_add

testgen_ad_aggregation

testgen_ad_avg_pool1d

testgen_ad_avg_pool2d

testgen_ad_broadcast

testgen_ad_cat

testgen_ad_complex

testgen_ad_conv1d

testgen_ad_conv2d

testgen_ad_conv_transpose1d

testgen_ad_conv_transpose2d

testgen_ad_cos

testgen_ad_cross_entropy_loss

testgen_ad_div

testgen_ad_erf

testgen_ad_exp

testgen_ad_gather_scatter

testgen_ad_gelu

testgen_ad_log

testgen_ad_log1p

testgen_ad_mask

testgen_ad_matmul

testgen_ad_max_pool1d

testgen_ad_max_pool2d

testgen_ad_maxmin

testgen_ad_mul

testgen_ad_multithread

testgen_ad_neg

testgen_ad_powf

testgen_ad_recip

testgen_ad_relu

testgen_ad_reshape

testgen_ad_select

testgen_ad_sin

testgen_ad_slice

testgen_ad_softmax

testgen_ad_sqrt

testgen_ad_sub

testgen_ad_tanh

testgen_ad_transpose

testgen_all

testgen_gradients

testgen_module_backward

Structs

Autodiff

Enable auto-differentiation on a backend.