Struct burn::backend::ndarray::NdArray

pub struct NdArray<E = f32> { /* private fields */ }

Tensor backend that uses the ndarray crate for executing tensor operations.

This backend is compatible with CPUs and can be compiled for almost any platform, including wasm, arm, and x86.