Struct burn::backend::LibTorch

pub struct LibTorch<E = f32> { /* private fields */ }

Tensor backend that uses LibTorch with the [tch] crate for executing tensor operations.

This backend is compatible with a wide range of hardwares ranging from CPUs to GPUs, but requires LibTorch to be installed correctly. The CPU version can be downloaded automatically and the CUDA version as well by setting the TORCH_CUDA_VERSION environment variable. For more complex configurations, check out the manual installation for burn-tch.

Refer to the [tch] crate for more information.