Struct burn::backend::tch::TchTensor

pub struct TchTensor<E, const D: usize>where
    E: Element,{
    pub tensor: Tensor,
    pub storage: Rc<*mut c_void, Global>,
    /* private fields */
}

A tensor that uses the tch backend.

Fields
tensor: Tensor

Handle to the tensor. Call methods on this field.

storage: Rc<*mut c_void, Global>

The tensor’s storage

Implementations

impl<E, const D: usize> TchTensor<E, D>where E: Element,

pub fn new(tensor: Tensor) -> TchTensor<E, D>

Create a new tensor.

Note that if the tensor was created from an operation that may reuse the same tensor storage as the parent, you should use from_existing instead.

pub fn from_existing( tensor: Tensor, storage_parent: Rc<*mut c_void, Global> ) -> TchTensor<E, D>

Create a tensor that was created from an operation executed on a parent tensor.

If the child tensor shared the same storage as its parent, it will be cloned, effectively tracking how much tensors point to the same memory space.

impl<P, const D: usize> TchTensor<P, D>where P: Element,

pub fn mut_ops<F, EOut, const D_OUT: usize>( &mut self, func: F ) -> Option<TchTensor<EOut, D_OUT>>where F: Fn(&mut Tensor) -> Tensor, EOut: Element,

Execute an operation on a tensor if the data can be reused.

pub fn unary_ops<FOwn, FRef, EOut, const D_OUT: usize>( self, fown: FOwn, fref: FRef ) -> TchTensor<EOut, D_OUT>where EOut: Element, FOwn: Fn(Tensor) -> Tensor, FRef: Fn(&Tensor) -> Tensor,

Execute a unary ops reusing the tensor data if possible.

pub fn binary_ops_tensor<FLMut, FRMut, FRef, EOut, const D_OUT: usize>( lhs: TchTensor<P, D>, rhs: TchTensor<P, D>, flmut: FLMut, frmut: FRMut, fref: FRef ) -> TchTensor<EOut, D_OUT>where EOut: Element, FLMut: Fn(&mut Tensor, &Tensor) -> Tensor, FRMut: Fn(&Tensor, &mut Tensor) -> Tensor, FRef: Fn(&Tensor, &Tensor) -> Tensor,

Execute a binary ops reusing the tensor data if possible.

impl<E, const D: usize> TchTensor<E, D>where E: Element + Default,

pub fn from_data(data: Data<E, D>, device: Device) -> TchTensor<E, D>

Creates a new tensor from a shape and a device.

Arguments

data - The tensor’s data.

device - The device on which the tensor will be allocated.

Returns

A new tensor.

impl<E, const D: usize> TchTensor<E, D>where E: Element + Default + Copy + Debug,

pub fn empty(shape: Shape<D>, device: TchDevice) -> TchTensor<E, D>

Creates an empty tensor from a shape and a device.

Arguments

shape - The shape of the tensor.

device - The device to create the tensor on.

Returns

A new empty tensor.