Module burn::backend::wgpu::compute

Compute related module.

Structs

DynamicKernel

Wraps a dynamic kernel source into a kernel with launch information such as workgroup.

StaticKernel

Wraps a static kernel source into a kernel with launch information such as workgroup.

WgpuResource

The memory resource that can be allocated for wgpu.

WgpuServer

Wgpu compute server.

WgpuStorage

Buffer storage for wgpu.

WorkGroup

Provides launch information specifying the number of work groups to be used by a compute shader.

Enums

WgpuResourceKind

How the resource is used, either as a slice or fully.

Traits

Kernel

Kernel trait with the source that will be compiled and cached based on the provided id.

Functions

compute_client

Get the compute client for the given device.

init_async

Init the client async, necessary for wasm.

select_device

Select the wgpu device and queue based on the provided device.

Type Definitions

WgpuComputeClient

Wgpu compute client to communicate with the compute server.

WgpuHandle

Wgpu server handle.