mabor::data::dataset

Module vision

Vision datasets.

Structs

BoundingBox

Object detection bounding box annotation.

ImageDatasetItem

Image dataset item.

ImageFolderDataset

A generic dataset to load images from disk.

MnistDataset

The MNIST dataset consists of 70,000 28x28 black-and-white images in 10 classes (one for each digits), with 7,000 images per class. There are 60,000 training images and 10,000 test images.

MnistItem

MNIST item.

SegmentationMask

Segmentation mask annotation. For semantic segmentation, a mask has a single channel (C = 1). For instance segmentation, there may be multiple masks per image (C >= 1).

Enums

Annotation

Annotation type for different tasks.

ImageLoaderError

Error type for ImageFolderDataset.

PixelDepth

Image data type.