mabor::module

Trait ModuleDisplayDefault

pub trait ModuleDisplayDefault {
    // Required method
    fn content(&self, _content: Content) -> Option<Content>;

    // Provided method
    fn num_params(&self) -> usize { ... }
}

Default display settings for a module.

Required Methods

fn content(&self, _content: Content) -> Option<Content>

Attributes of the module used for display purposes.

Arguments

_content - The content object that contains display settings and attributes.

Returns

An optional content object containing the display attributes.

Provided Methods

fn num_params(&self) -> usize

Gets the number of the parameters of the module.