Enum burn::nn::Initializer

pub enum Initializer {
    Constant {
        value: f64,
    },
    Ones,
    Zeros,
    Uniform {
        min: f64,
        max: f64,
    },
    Normal {
        mean: f64,
        std: f64,
    },
    KaimingUniform {
        gain: f64,
        fan_out_only: bool,
    },
    KaimingNormal {
        gain: f64,
        fan_out_only: bool,
    },
    XavierUniform {
        gain: f64,
    },
    XavierNormal {
        gain: f64,
    },
}

Enum specifying with what values a tensor should be initialized

Variants

Constant

Fields

value: f64

The value to fill the tensor with

Fills tensor with specified value everywhere

Ones

Fills tensor with 1s everywhere

Zeros

Fills tensor with 0s everywhere

Uniform

Fields

min: f64

The minimum value to draw from

max: f64

The maximum value to draw from

Fills tensor with values drawn uniformly between specified values

Normal

Fields

mean: f64

The mean of the normal distribution

std: f64

The standard deviation of the normal distribution

Fills tensor with values drawn from normal distribution with specified mean and std

KaimingUniform

Fields

gain: f64

The gain to use in initialization formula

fan_out_only: bool

Whether to use fan out only in initialization formula

Fills tensor with values according to the uniform version of Kaiming initialization

KaimingNormal

Fields

gain: f64

The gain to use in initialization formula

fan_out_only: bool

Whether to use fan out only in initialization formula

Fills tensor with values according to the uniform version of Kaiming initialization

XavierUniform

Fields

gain: f64

The gain to use in initialization formula

Fills tensor with values according to the uniform version of Xavier Glorot initialization described in Understanding the difficulty of training deep feedforward neural networks

XavierNormal

Fields

gain: f64

The gain to use in initialization formula

Fills tensor with values according to the normal version of Xavier Glorot initialization described in Understanding the difficulty of training deep feedforward neural networks

Implementations

impl Initializer

pub fn init<B, const D: usize, S>(&self, shape: S) -> Tensor<B, D, Float>where B: Backend, S: Into<Shape<D>>,

Inits a tensor of given shape with values depending on initializer kind.

Params

shape: Shape of the initiated tensor.

pub fn init_with<B, const D: usize, S>( &self, shape: S, fan_in: Option<usize>, fan_out: Option<usize> ) -> Tensor<B, D, Float>where B: Backend, S: Into<Shape<D>>,

Inits a tensor of given shape with values depending on initializer kind, with the possibility of specifying fan in and fan out

Params

shape: Shape of the initiated tensor.

fan_in: Option<usize>, the fan in to use in initialization formula, if needed

fan_out: Option<usize>, the fan out to use in initialization formula, if needed