mabor::nn

Module loss

Loss module

Structs

BinaryCrossEntropyLoss

Calculate the binary cross entropy loss from the input logits and the targets.

BinaryCrossEntropyLossConfig

Configuration to create a Binary Cross-entropy loss using the init function.

BinaryCrossEntropyLossRecord

The record type for the module.

BinaryCrossEntropyLossRecordItem

The record item type for the module.

CosineEmbeddingLoss

Cosine embedding loss between two tensors.

CosineEmbeddingLossConfig

Configuration for CosineEmbeddingLoss.

CrossEntropyLoss

Calculate the cross entropy loss from the input logits and the targets.

CrossEntropyLossConfig

Configuration to create a Cross-entropy loss using the init function.

CrossEntropyLossRecord

The record type for the module.

CrossEntropyLossRecordItem

The record item type for the module.

HuberLoss

Calculate the Huber loss between the inputs and the target.

HuberLossConfig

Configuration to create a Huber loss.

MseLoss

Calculate the mean squared error loss from the input logits and the targets.

PoissonNllLoss

Negative Log Likelihood (NLL) loss with a Poisson distribution assumption for the target.

PoissonNllLossConfig

Configuration for creating a PoissonNllLoss instance.

Enums

Reduction

The reduction type for the loss.