burn::nn

Module lstm

Long Short-Term Memory module.

Structs

BiLstm

The BiLstm module. This implementation is for Bidirectional LSTM.

BiLstmConfig

Configuration to create a BiLstm module using the init function.

BiLstmRecord

The record type for the module.

BiLstmRecordItem

The record item type for the module.

Lstm

The Lstm module. This implementation is for a unidirectional, stateless, Lstm.

LstmConfig

Configuration to create a Lstm module using the init function.

LstmRecord

The record type for the module.

LstmRecordItem

The record item type for the module.

LstmState

A LstmState is used to store cell state and hidden state in LSTM.