mabor::nn

Module transformer

Transformer module

Structs

PositionWiseFeedForward

Applies the position-wise feed-forward network to the input tensor from the paper Attention Is All You Need.

PositionWiseFeedForwardConfig

Configuration to create a position-wise feed-forward layer using the init function.

PositionWiseFeedForwardRecord

The record type for the module.

PositionWiseFeedForwardRecordItem

The record item type for the module.

TransformerDecoder

The transformer decoder module as describe in the paper Attention Is All You Need.

TransformerDecoderAutoregressiveCache

Autoregressive cache for the Transformer Decoder layer.

TransformerDecoderConfig

Configuration to create a Transformer Decoder layer using the init function.

TransformerDecoderInput

Transformer Decoder forward pass input argument.

TransformerDecoderLayer

Transformer Decoder layer module.

TransformerDecoderLayerRecord

The record type for the module.

TransformerDecoderLayerRecordItem

The record item type for the module.

TransformerDecoderRecord

The record type for the module.

TransformerDecoderRecordItem

The record item type for the module.

TransformerEncoder

The transformer encoder module as describe in the paper Attention Is All You Need.

TransformerEncoderAutoregressiveCache

Autoregressive cache for the Transformer Encoder layer.

TransformerEncoderConfig

Configuration to create a Transformer Encoder layer using the init function.

TransformerEncoderInput

Transformer Encoder forward pass input argument.

TransformerEncoderLayer

Transformer encoder layer module.

TransformerEncoderLayerRecord

The record type for the module.

TransformerEncoderLayerRecordItem

The record item type for the module.

TransformerEncoderRecord

The record type for the module.

TransformerEncoderRecordItem

The record item type for the module.