mabor

Module optim

Optimizer module.

Modules

adaptor

Adaptor module for optimizers.

decay

Weight decay module for optimizers.

momentum

Momentum module for optimizers.

record

Record module for optimizers.

Structs

AdaGrad

AdaGrad optimizer

AdaGradConfig

AdaGrad configuration.

AdaGradState

AdaGrad state.

AdaGradStateItem

The record item type for the module.

Adam

Adam optimizer as described in the paper Adam: A Method for Stochastic Optimization.

AdamConfig

Adam configuration.

AdamState

Adam state.

AdamStateItem

The record item type for the module.

AdamW

AdamW optimizer as described in the paper Decoupled Weight Decay Regularization, Loshchilov and Hutter, 2019.

AdamWConfig

AdamW configuration.

AdamWState

AdamW state.

AdamWStateItem

The record item type for the module.

AdaptiveMomentumState

Adaptive momentum state.

AdaptiveMomentumStateItem

The record item type for the module.

CenteredState

CenteredState is to store and pass optimizer step params.

CenteredStateItem

The record item type for the module.

GradientsAccumulator

Accumulate gradients into a single Gradients object.

GradientsParams

Data type that contains gradients for parameters.

LrDecayState

Learning rate decay state (also includes sum state).

LrDecayStateItem

The record item type for the module.

RmsProp

Optimizer that implements stochastic gradient descent with momentum. The optimizer can be configured with RmsPropConfig.

RmsPropConfig

Configuration to create the RmsProp optimizer.

RmsPropMomentum

RmsPropMomentum is to store config status for optimizer. (, which is stored in optimizer itself and not passed in during step() calculation)

RmsPropMomentumState

RmsPropMomentumState is to store and pass optimizer step params.

RmsPropMomentumStateItem

The record item type for the module.

RmsPropState

State of RmsProp

RmsPropStateItem

The record item type for the module.

Sgd

Optimizer that implements stochastic gradient descent with momentum.

SgdConfig

Configuration to create the Sgd optimizer.

SgdState

State of Sgd.

SgdStateItem

The record item type for the module.

SquareAvgState

SquareAvgState is to store and pass optimizer step params.

SquareAvgStateItem

The record item type for the module.

Traits

Optimizer

General trait to optimize module.

SimpleOptimizer

Simple optimizer is an opinionated trait to simplify the process of implementing an optimizer.