burn

Module prelude

Structs and macros used by most projects. Add use burn::prelude::* to your code to quickly get started with Burn.

Macros

s

Creates a slice specification for tensor indexing operations.

Structs

Bool

A type-level representation of the kind of a bool tensor.

Float

A type-level representation of the kind of a float tensor

Int

A type-level representation of the kind of a int tensor.

Shape

Shape of a tensor.

Tensor

A tensor with a given backend, shape and data type.

TensorData

Data structure for tensors.

Traits

Backend

This trait defines all types and functions needed for a backend to be used with mabor.

Config

Configuration trait.

ElementConversion

Element conversion trait for tensor.

Module

Trait for all neural network modules.

Type Aliases

Device

Device type used by the backend.

Derive Macros

Config

Derive macro for the config.

Module

Derive macro for the module.