mabor

Module record

Module for the recorder.

Structs

BinBytesRecorder

In memory recorder using the bincode format.

BinFileRecorder

File recorder using the bincode format.

BinGzFileRecorder

File recorder using the bincode format compressed with gzip.

BurnMetadata

Metadata of a record.

BurnRecord

Record that can be saved by a Recorder.

BurnRecordNoItem

Record that can be saved by a Recorder without the item.

DoublePrecisionSettings

Precision settings optimized for precision.

FullPrecisionSettings

Default precision settings.

HalfPrecisionSettings

Precision settings optimized for compactness.

JsonGzFileRecorder

File recorder using the json format compressed with gzip.

NamedMpkBytesRecorder

In memory recorder using the Named MessagePack.

NamedMpkFileRecorder

File recorder using the named msgpack format.

NamedMpkGzFileRecorder

File recorder using the named msgpack format compressed with gzip.

ParamSerde

(De)serialize parameters into a clean format.

PrettyJsonFileRecorder

File recorder using pretty json format for easy readability.

Enums

RecorderError

Error that can occur when using a Recorder.

Traits

BytesRecorder

Recorder trait specialized to save and load data to and from bytes.

FileRecorder

Recorder trait specialized to save and load data to and from files.

PrecisionSettings

Settings allowing to control the precision when (de)serializing items.

Record

Trait to define a family of types which can be recorded using any settings.

Recorder

Record any item implementing Serialize and DeserializeOwned.

Type Aliases

CompactRecorder

Recorder optimized for compactness.

DebugRecordSettings

Debug recorder.

DefaultFileRecorder

Default file recorder.

DefaultRecorder

Default recorder.

NoStdInferenceRecorder

Inference recorder compatible with no-std.

NoStdTrainingRecorder

Training recorder compatible with no-std inference.

SensitiveCompactRecorder

Recorder optimized for compactness making it a good choice for model deployment.

Derive Macros

Record

Derive macro for the record.