Enum burn::record::serde::data::NestedValue

pub enum NestedValue {
    Default,
    Bool(bool),
    String(String),
    F32(f32),
    F64(f64),
    I16(i16),
    I32(i32),
    I64(i64),
    U16(u16),
    U64(u64),
    Map(HashMap<String, NestedValue>),
    Vec(Vec<NestedValue>),
}

The main data structure used for deserialization.

It can hold tree-like structures of nested maps and vectors.

Variants

Default

The default value, which actually does not hold any value and it is used to indicate that the value should be populated with the default value.

Bool(bool)

A boolean value.

String(String)

A string value.

F32(f32)

Floating point 32-bit value.

F64(f64)

Floating point 64-bit value.

I16(i16)

Signed 16-bit integer value.

I32(i32)

Signed 32-bit integer value.

I64(i64)

Signed 64-bit integer value.

U16(u16)

Unsigned 16-bit integer value used for bf16 and f16 serialization

U64(u64)

Unsigned 64-bit integer value.

Map(HashMap<String, NestedValue>)

A map of nested values (typically used for structs)

Vec(Vec<NestedValue>)

A vector of nested values (typically used for vector of structs or numbers)

Implementations

impl NestedValue

pub fn as_map(self) -> Option<HashMap<String, NestedValue>>

Get the nested value as a map.

pub fn as_bool(self) -> Option<bool>

Get the nested value as a boolean.

pub fn as_string(self) -> Option<String>

Get the nested value as a string.

pub fn as_f32(self) -> Option<f32>

Get the nested value as a f32.

pub fn as_f64(self) -> Option<f64>

Get the nested value as a f64.

pub fn as_i16(self) -> Option<i16>

Get the nested value as an i16.

pub fn as_i32(self) -> Option<i32>

Get the nested value as an i32.

pub fn as_i64(self) -> Option<i64>

Get the nested value as an i64.

pub fn as_u16(self) -> Option<u16>

Get the nested value as a u16.

pub fn as_u64(self) -> Option<u64>

Get the nested value as a u64.

pub fn try_into_record<T, PS, A, B>( self, device: &<B as Backend>::Device ) -> Result<T, Error>where B: Backend, T: Record<B>, PS: PrecisionSettings, A: BurnModuleAdapter,

Deserialize a nested value into a record type.