mabor::record

Trait BytesRecorder

pub trait BytesRecorder<B, L>: Recorder<B, RecordArgs = (), RecordOutput = Vec<u8>, LoadArgs = L>
where
    B: Backend,
    L: AsRef<[u8]> + Send + Sync + Debug + Clone + Default,
{ }

Recorder trait specialized to save and load data to and from bytes.

Notes

This is especially useful in no_std environment where weights are stored directly in compiled binaries.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.