mabor::record

Trait FileRecorder

pub trait FileRecorder<B>: Recorder<B, RecordArgs = PathBuf, RecordOutput = (), LoadArgs = PathBuf>
where
    B: Backend,
{
    // Required method
    fn file_extension() -> &'static str;
}

Recorder trait specialized to save and load data to and from files.

Required Methods

fn file_extension() -> &'static str

File extension of the format used by the recorder.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.