mabor::tensor::cast

Trait ToElement

pub trait ToElement {

    // Required methods
    fn to_i64(&self) -> i64;
    fn to_u64(&self) -> u64;

    // Provided methods
    fn to_isize(&self) -> isize { ... }
    fn to_i8(&self) -> i8 { ... }
    fn to_i16(&self) -> i16 { ... }
    fn to_i32(&self) -> i32 { ... }
    fn to_i128(&self) -> i128 { ... }
    fn to_usize(&self) -> usize { ... }
    fn to_u8(&self) -> u8 { ... }
    fn to_u16(&self) -> u16 { ... }
    fn to_u32(&self) -> u32 { ... }
    fn to_u128(&self) -> u128 { ... }
    fn to_f16(&self) -> f16 { ... }
    fn to_bf16(&self) -> bf16 { ... }
    fn to_f32(&self) -> f32 { ... }
    fn to_f64(&self) -> f64 { ... }
    fn to_bool(&self) -> bool { ... }

}

A generic trait for converting a value to a number. Adapted from num_traits::ToPrimitive to support bool.

A value can be represented by the target type when it lies within the range of scalars supported by the target type. For example, a negative integer cannot be represented by an unsigned integer type, and an i64 with a very high magnitude might not be convertible to an i32. On the other hand, conversions with possible precision loss or truncation are admitted, like an f32 with a decimal part to an integer type, or even a large f64 saturating to f32 infinity.

The methods panic when the value cannot be represented by the target type.

Required Methods

fn to_i64(&self) -> i64

Converts the value of self to an i64.

fn to_u64(&self) -> u64

Converts the value of self to a u64.

Provided Methods

fn to_isize(&self) -> isize

Converts the value of self to an isize.

fn to_i8(&self) -> i8

Converts the value of self to an i8.

fn to_i16(&self) -> i16

Converts the value of self to an i16.

fn to_i32(&self) -> i32

Converts the value of self to an i32.

fn to_i128(&self) -> i128

Converts the value of self to an i128.

The default implementation converts through to_i64(). Types implementing this trait should override this method if they can represent a greater range.

fn to_usize(&self) -> usize

Converts the value of self to a usize.

fn to_u8(&self) -> u8

Converts the value of self to a u8.

fn to_u16(&self) -> u16

Converts the value of self to a u16.

fn to_u32(&self) -> u32

Converts the value of self to a u32.

fn to_u128(&self) -> u128

Converts the value of self to a u128.

The default implementation converts through to_u64(). Types implementing this trait should override this method if they can represent a greater range.

fn to_f16(&self) -> f16

Converts the value of self to an f16. Overflows may map to positive or negative infinity.

fn to_bf16(&self) -> bf16

Converts the value of self to an bf16. Overflows may map to positive or negative infinity.

fn to_f32(&self) -> f32

Converts the value of self to an f32. Overflows may map to positive or negative infinity.

fn to_f64(&self) -> f64

Converts the value of self to an f64. Overflows may map to positive or negative infinity.

The default implementation tries to convert through to_i64(), and failing that through to_u64(). Types implementing this trait should override this method if they can represent a greater range.

fn to_bool(&self) -> bool

Converts the value of self to a bool. Rust only considers 0 and 1 to be valid booleans, but for compatibility, C semantics are adopted (anything that’s not 0 is true).

The default implementation tries to convert through to_i64(), and failing that through to_u64(). Types implementing this trait should override this method if they can represent a greater range.