mabor::tensor

Module grid

The grid module.

Structs

GridOptions

Grid policy options.

Enums

GridIndexing

Enum to specify index cardinal layout.

GridSparsity

Enum to specify grid sparsity mode.

IndexPos

Enum to specify the index dimension position.

Functions

meshgrid

Return a collection of coordinate matrices for coordinate vectors.

meshgrid_stack

Return a coordinate matrix for a given set of 1D coordinate tensors.