mabor::tensor

Module linalg

The linalg module.

Enums

Norm

Specifies the type of norm to compute.

Constants

DEFAULT_EPSILON

Default epsilon value to avoid division by zero

Functions

cosine_similarity

Computes the cosine similarity between two tensors along a specified dimension.

l0_norm

Computes the L0 norm of a tensor along a specified dimension.

l1_norm

Computes the L1 norm of a tensor along a specified dimension.

l2_norm

Computes the L2 norm of a tensor along a specified dimension.

lp_norm

Computes the general L(p) norm of a tensor along a specified dimension.

max_abs_norm

Computes the L:INFINITY norm of a tensor along a specified dimension.

min_abs_norm

Computes the L:NEG_INFINITY norm of a tensor along a specified dimension.

vector_norm

Computes the vector norm of a tensor along a specified dimension.

vector_normalize

Normalize a tensor versus its vector_norm.