mabor::tensor::ops

Trait IntTensorOps

pub trait IntTensorOps<B>
where
    B: Backend,
{

    // Required methods
    fn int_empty(
        shape: Shape,
        device: &<B as Backend>::Device,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_into_data(
        tensor: <B as Backend>::IntTensorPrimitive,
    ) -> impl Future<Output = TensorData> + Send;
    fn int_from_data(
        data: TensorData,
        device: &<B as Backend>::Device,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_device(
        tensor: &<B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::Device;
    fn int_to_device(
        tensor: <B as Backend>::IntTensorPrimitive,
        device: &<B as Backend>::Device,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_reshape(
        tensor: <B as Backend>::IntTensorPrimitive,
        shape: Shape,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_slice(
        tensor: <B as Backend>::IntTensorPrimitive,
        indices: &[Range<usize>],
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_slice_assign(
        tensor: <B as Backend>::IntTensorPrimitive,
        indices: &[Range<usize>],
        value: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_into_float(
        tensor: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::FloatTensorPrimitive;
    fn int_mask_where(
        tensor: <B as Backend>::IntTensorPrimitive,
        mask: <B as Backend>::BoolTensorPrimitive,
        source: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_mask_fill(
        tensor: <B as Backend>::IntTensorPrimitive,
        mask: <B as Backend>::BoolTensorPrimitive,
        value: <B as Backend>::IntElem,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_gather(
        dim: usize,
        tensor: <B as Backend>::IntTensorPrimitive,
        indices: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_scatter(
        dim: usize,
        tensor: <B as Backend>::IntTensorPrimitive,
        indices: <B as Backend>::IntTensorPrimitive,
        value: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_select(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
        indices: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_select_assign(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
        indices: <B as Backend>::IntTensorPrimitive,
        value: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_equal(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::BoolTensorPrimitive;
    fn int_equal_elem(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::BoolTensorPrimitive;
    fn int_greater(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::BoolTensorPrimitive;
    fn int_greater_elem(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::BoolTensorPrimitive;
    fn int_greater_equal(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::BoolTensorPrimitive;
    fn int_greater_equal_elem(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::BoolTensorPrimitive;
    fn int_lower(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::BoolTensorPrimitive;
    fn int_lower_elem(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::BoolTensorPrimitive;
    fn int_lower_equal(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::BoolTensorPrimitive;
    fn int_lower_equal_elem(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::BoolTensorPrimitive;
    fn int_add(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_add_scalar(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_sub(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_sub_scalar(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_mul(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_mul_scalar(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_div(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_div_scalar(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_remainder(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_remainder_scalar(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_zeros(
        shape: Shape,
        device: &<B as Backend>::Device,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_ones(
        shape: Shape,
        device: &<B as Backend>::Device,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_sum(
        tensor: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_sum_dim(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_prod(
        tensor: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_prod_dim(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_mean_dim(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_argmax(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_argmin(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_abs(
        tensor: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_swap_dims(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim1: usize,
        dim2: usize,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_permute(
        tensor: <B as Backend>::IntTensorPrimitive,
        axes: &[usize],
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_flip(
        tensor: <B as Backend>::IntTensorPrimitive,
        axes: &[usize],
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_random(
        shape: Shape,
        distribution: Distribution,
        device: &<B as Backend>::Device,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn int_expand(
        tensor: <B as Backend>::IntTensorPrimitive,
        shape: Shape,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn bitwise_and(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn bitwise_and_scalar(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn bitwise_or(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn bitwise_or_scalar(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn bitwise_xor(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn bitwise_xor_scalar(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn bitwise_not(
        tensor: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn bitwise_left_shift(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn bitwise_left_shift_scalar(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn bitwise_right_shift(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive;
    fn bitwise_right_shift_scalar(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::IntTensorPrimitive;

    // Provided methods
    fn int_repeat_dim(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
        times: usize,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_cat(
        tensors: Vec<<B as Backend>::IntTensorPrimitive>,
        dim: usize,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_not_equal(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::BoolTensorPrimitive { ... }
    fn int_not_equal_elem(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::BoolTensorPrimitive { ... }
    fn int_powi(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_powf(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::FloatTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_powi_scalar(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: <B as Backend>::IntElem,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_powf_scalar(
        lhs: <B as Backend>::IntTensorPrimitive,
        rhs: f32,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_clamp_min(
        tensor: <B as Backend>::IntTensorPrimitive,
        min: <B as Backend>::IntElem,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_clamp_max(
        tensor: <B as Backend>::IntTensorPrimitive,
        max: <B as Backend>::IntElem,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_clamp(
        tensor: <B as Backend>::IntTensorPrimitive,
        min: <B as Backend>::IntElem,
        max: <B as Backend>::IntElem,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_neg(
        tensor: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_full(
        shape: Shape,
        fill_value: <B as Backend>::IntElem,
        device: &<B as Backend>::Device,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_mean(
        tensor: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_max(
        tensor: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_max_dim(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_max_dim_with_indices(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
    ) -> (<B as Backend>::IntTensorPrimitive, <B as Backend>::IntTensorPrimitive) { ... }
    fn int_max_abs(
        tensor: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_max_abs_dim(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_min(
        tensor: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_min_dim(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_min_dim_with_indices(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
    ) -> (<B as Backend>::IntTensorPrimitive, <B as Backend>::IntTensorPrimitive) { ... }
    fn int_transpose(
        tensor: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_arange_step(
        range: Range<i64>,
        step: usize,
        device: &<B as Backend>::Device,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_arange(
        range: Range<i64>,
        device: &<B as Backend>::Device,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_any(
        tensor: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::BoolTensorPrimitive { ... }
    fn int_any_dim(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
    ) -> <B as Backend>::BoolTensorPrimitive { ... }
    fn int_all(
        tensor: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::BoolTensorPrimitive { ... }
    fn int_all_dim(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
    ) -> <B as Backend>::BoolTensorPrimitive { ... }
    fn int_sign(
        tensor: <B as Backend>::IntTensorPrimitive,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_sort(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
        descending: bool,
    ) -> <B as Backend>::IntTensorPrimitive { ... }
    fn int_sort_with_indices(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
        descending: bool,
    ) -> (<B as Backend>::IntTensorPrimitive, <B as Backend>::IntTensorPrimitive) { ... }
    fn int_argsort(
        tensor: <B as Backend>::IntTensorPrimitive,
        dim: usize,
        descending: bool,
    ) -> <B as Backend>::IntTensorPrimitive { ... }

}

Int Tensor API for basic and numeric operations, see tensor for documentation on each function.

Required Methods

fn int_empty( shape: Shape, device: &<B as Backend>::Device, ) -> <B as Backend>::IntTensorPrimitive

Creates a new int tensor.

Arguments

shape - The shape of the tensor.

device - The device to create the tensor on.

Returns

The integer tensor with the given shape.

fn int_into_data( tensor: <B as Backend>::IntTensorPrimitive, ) -> impl Future<Output = TensorData> + Send

Converts the tensor to a data structure.

Arguments

tensor - The tensor.

Returns

The data structure with the tensor’s data.

fn int_from_data( data: TensorData, device: &<B as Backend>::Device, ) -> <B as Backend>::IntTensorPrimitive

Creates a tensor from the data structure.

Arguments

data - The data structure.

device - The device to create the tensor on.

Returns

The tensor with the data.

fn int_device( tensor: &<B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::Device

Gets the device of the tensor.

Arguments

tensor - The tensor.

Returns

The device of the tensor.

fn int_to_device( tensor: <B as Backend>::IntTensorPrimitive, device: &<B as Backend>::Device, ) -> <B as Backend>::IntTensorPrimitive

Moves the tensor to the given device.

fn int_reshape( tensor: <B as Backend>::IntTensorPrimitive, shape: Shape, ) -> <B as Backend>::IntTensorPrimitive

Reshapes the tensor.

Arguments

tensor - The tensor.

shape - The new shape.

Returns

The tensor with the new shape.

fn int_slice( tensor: <B as Backend>::IntTensorPrimitive, indices: &[Range<usize>], ) -> <B as Backend>::IntTensorPrimitive

Gets the element at the given indices.

Arguments

tensor - The tensor.

indices - The indices.

Returns

The elements at the given indices.

fn int_slice_assign( tensor: <B as Backend>::IntTensorPrimitive, indices: &[Range<usize>], value: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Sets the element at the given indices.

Arguments

tensor - The tensor.

indices - The indices.

Returns

The tensor with the element at the given indices set.

fn int_into_float( tensor: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::FloatTensorPrimitive

Converts int tensor to float tensor.

Arguments

tensor - The tensor.

Returns

The int tensor with the same data as the float tensor.

fn int_mask_where( tensor: <B as Backend>::IntTensorPrimitive, mask: <B as Backend>::BoolTensorPrimitive, source: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Fills the tensor with values from the source tensor if the mask is true at the given indices.

Arguments

tensor - The tensor.

mask - The mask.

source - The source tensor.

Returns

The tensor with the values filled.

fn int_mask_fill( tensor: <B as Backend>::IntTensorPrimitive, mask: <B as Backend>::BoolTensorPrimitive, value: <B as Backend>::IntElem, ) -> <B as Backend>::IntTensorPrimitive

Fills the tensor with the given value if the mask is true at the given indices.

Arguments

tensor - The tensor.

mask - The mask.

value - The value.

Returns

The tensor with the values filled.

fn int_gather( dim: usize, tensor: <B as Backend>::IntTensorPrimitive, indices: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Gather elements from the tensor at the given indices.

Arguments

dim - The dimension to gather from.

tensor - The tensor.

indices - The indices.

fn int_scatter( dim: usize, tensor: <B as Backend>::IntTensorPrimitive, indices: <B as Backend>::IntTensorPrimitive, value: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Scatter a given value to the tensor at the given indices.

Arguments

dim - The dimension to scatter to.

tensor - The tensor.

indices - The indices.

value - The value.

Returns

The tensor with the values scattered.

fn int_select( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, indices: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Select tensor elements along the given dimension corresponding to the given indices.

Arguments

tensor - The tensor.

dim - The dimension to select from.

indices - The indices.

Returns

The tensor with the selected elements.

fn int_select_assign( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, indices: <B as Backend>::IntTensorPrimitive, value: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Assign the selected elements along the given dimension corresponding to the given indices to the given value.

Arguments

tensor - The tensor.

dim - The dimension to select from.

indices - The indices.

value - The value.

Returns

The tensor with the selected elements assigned to the given value.

fn int_equal( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::BoolTensorPrimitive

Element-wise equality comparison.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

The boolean tensor with the result of the comparison.

fn int_equal_elem( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::BoolTensorPrimitive

Element-wise equality comparison with a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The boolean tensor with the result of the comparison.

fn int_greater( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::BoolTensorPrimitive

Element-wise greater than comparison.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

The boolean tensor with the result of the comparison.

fn int_greater_elem( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::BoolTensorPrimitive

Element-wise greater than comparison with a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The boolean tensor with the result of the comparison.

fn int_greater_equal( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::BoolTensorPrimitive

Element-wise greater than or equal comparison.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

The boolean tensor with the result of the comparison.

fn int_greater_equal_elem( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::BoolTensorPrimitive

Element-wise greater than or equal comparison with a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The boolean tensor with the result of the comparison.

fn int_lower( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::BoolTensorPrimitive

Element-wise less than comparison.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

The boolean tensor with the result of the comparison.

fn int_lower_elem( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::BoolTensorPrimitive

Element-wise less than comparison with a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The boolean tensor with the result of the comparison.

fn int_lower_equal( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::BoolTensorPrimitive

Element-wise less than or equal comparison.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

The boolean tensor with the result of the comparison.

fn int_lower_equal_elem( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::BoolTensorPrimitive

Element-wise less than or equal comparison with a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The boolean tensor with the result of the comparison.

fn int_add( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Element-wise addition.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

The result of the addition.

fn int_add_scalar( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::IntTensorPrimitive

Element-wise addition with a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The result of the addition.

fn int_sub( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Element-wise subtraction.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

The result of the subtraction.

fn int_sub_scalar( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::IntTensorPrimitive

Element-wise subtraction with a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The result of the subtraction.

fn int_mul( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Element-wise multiplication.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

The result of the multiplication.

fn int_mul_scalar( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::IntTensorPrimitive

Element-wise multiplication with a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The result of the multiplication.

fn int_div( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Element-wise division.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

The result of the division.

fn int_div_scalar( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::IntTensorPrimitive

Element-wise division with a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The result of the division.

fn int_remainder( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Element-wise modulus.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The result of applying the modulus of the scalar to the tensor.

fn int_remainder_scalar( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::IntTensorPrimitive

Element-wise modulus with a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The result of applying the modulus of the scalar to the tensor.

fn int_zeros( shape: Shape, device: &<B as Backend>::Device, ) -> <B as Backend>::IntTensorPrimitive

Creates a tensor of zeros.

Arguments

shape - The shape of the tensor.

device - The device to create the tensor on.

Returns

The tensor of zeros.

fn int_ones( shape: Shape, device: &<B as Backend>::Device, ) -> <B as Backend>::IntTensorPrimitive

Creates a tensor of ones.

Arguments

shape - The shape of the tensor.

device - The device to create the tensor on.

Returns

The tensor of ones.

fn int_sum( tensor: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Sums all elements in the tensor.

Arguments

tensor - The tensor to sum.

Returns

The sum of all elements in the tensor.

fn int_sum_dim( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, ) -> <B as Backend>::IntTensorPrimitive

Sums all elements in the tensor along a dimension.

Arguments

tensor - The tensor to sum.

dim - The dimension to sum along.

Returns

The sum of all elements in the tensor along the dimension.

fn int_prod( tensor: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Computes the product of all elements in the tensor.

Arguments

tensor - The tensor to compute the product of.

Returns

The product of all elements in the tensor.

fn int_prod_dim( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, ) -> <B as Backend>::IntTensorPrimitive

Computes the product of all elements in the tensor along a dimension.

Arguments

tensor - The tensor to compute the product of.

dim - The dimension to compute the product along.

Returns

The product of all elements in the tensor along the dimension.

fn int_mean_dim( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, ) -> <B as Backend>::IntTensorPrimitive

Computes the mean of all elements in the tensor along a dimension.

Arguments

tensor - The tensor to compute the mean of.

Returns

The mean of all elements in the tensor along the dimension.

fn int_argmax( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, ) -> <B as Backend>::IntTensorPrimitive

Gets the indices of the maximum elements along a dimension.

Arguments

tensor - The tensor to get the maximum indices of.

dim - The dimension to get the maximum indices along.

Returns

The indices of the maximum elements along the dimension.

fn int_argmin( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, ) -> <B as Backend>::IntTensorPrimitive

Gets the indices of the minimum elements along a dimension.

Arguments

tensor - The tensor to get the minimum indices of.

dim - The dimension to get the minimum indices along.

Returns

The indices of the minimum elements along the dimension.

fn int_abs( tensor: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Returns a new tensor with absolute values.

Arguments

tensor - The tensor to take absolute value of.

Returns

A tensor with the same shape as tensor with absolute values.

fn int_swap_dims( tensor: <B as Backend>::IntTensorPrimitive, dim1: usize, dim2: usize, ) -> <B as Backend>::IntTensorPrimitive

Swaps two dimensions of an int tensor.

Arguments

tensor - The tensor to swap the dimensions of.

dim1 - The first dimension to swap.

dim2 - The second dimension to swap.

Returns

The tensor with the dimensions swapped.

fn int_permute( tensor: <B as Backend>::IntTensorPrimitive, axes: &[usize], ) -> <B as Backend>::IntTensorPrimitive

Permutes the dimensions of a tensor.

Arguments

tensor - The tensor to permute the dimensions of.

axes - The new order of the dimensions.

Returns

The tensor with the dimensions permuted.

fn int_flip( tensor: <B as Backend>::IntTensorPrimitive, axes: &[usize], ) -> <B as Backend>::IntTensorPrimitive

Reverse the order of elements in a tensor along the given axes.

Arguments

tensor - The tensor to reverse.

axes - The axes to reverse.

The tensor with the elements reversed.

fn int_random( shape: Shape, distribution: Distribution, device: &<B as Backend>::Device, ) -> <B as Backend>::IntTensorPrimitive

Creates a new int tensor with random values.

Arguments

shape - The shape of the tensor.

distribution - The distribution to sample from.

device - The device to create the tensor on.

Returns

The tensor with the given shape and random values.

fn int_expand( tensor: <B as Backend>::IntTensorPrimitive, shape: Shape, ) -> <B as Backend>::IntTensorPrimitive

Broadcasts the int tensor to the given shape.

fn bitwise_and( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Bitwise AND operation for Int Tensors

fn bitwise_and_scalar( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::IntTensorPrimitive

Bitwise AND operation for Int Tensors with a scalar

fn bitwise_or( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Bitwise OR operation for Int Tensors

fn bitwise_or_scalar( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::IntTensorPrimitive

Bitwise OR operation for Int Tensors with a scalar

fn bitwise_xor( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Bitwise XOR operation for Int Tensors

fn bitwise_xor_scalar( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::IntTensorPrimitive

Bitwise XOR operation for Int Tensors with a scalar

fn bitwise_not( tensor: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Bitwise NOT operation for Int Tensors

fn bitwise_left_shift( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Bitwise left shift operation for Int Tensors

fn bitwise_left_shift_scalar( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::IntTensorPrimitive

Bitwise left shift operation for Int Tensors with a scalar

fn bitwise_right_shift( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Bitwise right shift operation for Int Tensors

fn bitwise_right_shift_scalar( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::IntTensorPrimitive

Bitwise right shift operation for Int Tensors with a scalar

Provided Methods

fn int_repeat_dim( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, times: usize, ) -> <B as Backend>::IntTensorPrimitive

Repeats the tensor along the given dimension the given number of times.

Arguments

tensor - The tensor.

dim - The dimension to repeat.

times - The number of times to repeat.

Returns

The tensor with the given dimension repeated the given number of times.

fn int_cat( tensors: Vec<<B as Backend>::IntTensorPrimitive>, dim: usize, ) -> <B as Backend>::IntTensorPrimitive

Concatenates the given tensors along the given dimension.

Arguments

tensors - The tensors.

dim - The dimension to concatenate along.

Returns

The concatenated tensor.

fn int_not_equal( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::BoolTensorPrimitive

Element-wise non-equality comparison.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

The boolean tensor with the result of the comparison.

fn int_not_equal_elem( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::BoolTensorPrimitive

Element-wise non-equality comparison with a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The boolean tensor with the result of the comparison.

fn int_powi( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Element-wise power with a IntTensor.

Arguments

lhs - The left hand side IntTensor.

rhs - The right hand side IntTensor.

Returns

The elements of lhs raised to the power of the elements of rhs.

fn int_powf( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::FloatTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Element-wise power with a floatTensor.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side floatTensor.

Returns

The elements of lhs raised to the value of rhs. Result is an IntTensor.

fn int_powi_scalar( lhs: <B as Backend>::IntTensorPrimitive, rhs: <B as Backend>::IntElem, ) -> <B as Backend>::IntTensorPrimitive

Element-wise power with a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The elements of lhs raised to the value of rhs.

fn int_powf_scalar( lhs: <B as Backend>::IntTensorPrimitive, rhs: f32, ) -> <B as Backend>::IntTensorPrimitive

Element-wise power with a floatTensor.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The elements of lhs raised to the value of rhs. Result is an IntTensor.

fn int_clamp_min( tensor: <B as Backend>::IntTensorPrimitive, min: <B as Backend>::IntElem, ) -> <B as Backend>::IntTensorPrimitive

Clamps a tensor under a minimum value.

Arguments

tensor - The tensor to clamp.

min - The minimum value.

Returns

The clamped tensor.

fn int_clamp_max( tensor: <B as Backend>::IntTensorPrimitive, max: <B as Backend>::IntElem, ) -> <B as Backend>::IntTensorPrimitive

Clamps a tensor over a maximum value.

Arguments

tensor - The tensor to clamp.

max - The maximum value.

Returns

The clamped tensor.

fn int_clamp( tensor: <B as Backend>::IntTensorPrimitive, min: <B as Backend>::IntElem, max: <B as Backend>::IntElem, ) -> <B as Backend>::IntTensorPrimitive

Clamps a tensor between a minimum and maximum value.

Arguments

tensor - The tensor to clamp.

min - The minimum value.

max - The maximum value.

Returns

The clamped tensor.

fn int_neg( tensor: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Element-wise negation.

Arguments

tensor - The tensor to negate.

Returns

The negated tensor.

fn int_full( shape: Shape, fill_value: <B as Backend>::IntElem, device: &<B as Backend>::Device, ) -> <B as Backend>::IntTensorPrimitive

Creates a tensor filled with given value.

Arguments

shape - The shape of the tensor.

fill_value - The value with which to fill the tensor.

device - The device to create the tensor on.

Returns

The tensor filled with given value

fn int_mean( tensor: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Computes the mean of all elements in the tensor.

Arguments

tensor - The tensor to compute the mean of.

Returns

The mean of all elements in the tensor.

fn int_max( tensor: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Gets the maximum element in the tensor.

Arguments

tensor - The tensor to get the maximum element of.

Returns

The maximum element in the tensor.

fn int_max_dim( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, ) -> <B as Backend>::IntTensorPrimitive

Gets the maximum element in the tensor along a dimension.

Arguments

tensor - The tensor to get the maximum element of.

dim - The dimension to get the maximum element along.

Returns

The maximum element in the tensor along the dimension.

fn int_max_dim_with_indices( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, ) -> (<B as Backend>::IntTensorPrimitive, <B as Backend>::IntTensorPrimitive)

Gets the maximum elements and corresponding indices along a dimension.

Arguments

tensor - The tensor to get the maximum elements and indices of.

dim - The dimension to get the maximum elements and indices along.

Returns

The maximum elements and corresponding indices along the dimension.

fn int_max_abs( tensor: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Gets the maximum absolute element in the tensor.

Arguments

tensor - The tensor to get the maximum element of.

Returns

The maximum element in the tensor.

fn int_max_abs_dim( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, ) -> <B as Backend>::IntTensorPrimitive

Gets the maximum absolute element in the tensor along a dimension.

Arguments

tensor - The tensor to get the maximum element of.

dim - The dimension to get the maximum element along.

Returns

The maximum element in the tensor along the dimension.

fn int_min( tensor: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Gets the minimum element in the tensor.

Arguments

tensor - The tensor to get the minimum element of.

Returns

The minimum element in the tensor.

fn int_min_dim( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, ) -> <B as Backend>::IntTensorPrimitive

Gets the minimum elements in the tensor along a dimension.

Arguments

tensor - The tensor to get the minimum element of.

dim - The dimension to get the minimum element along.

Returns

The minimum element in the tensor along the dimension.

fn int_min_dim_with_indices( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, ) -> (<B as Backend>::IntTensorPrimitive, <B as Backend>::IntTensorPrimitive)

Gets the minimum elements and corresponding indices along a dimension.

Arguments

tensor - The tensor to get the minimum elements and indices of.

dim - The dimension to get the minimum elements and indices along.

Returns

The minimum elements and corresponding indices along the dimension.

fn int_transpose( tensor: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Transposes an int tensor.

Arguments

tensor - The tensor to transpose.

Returns

The transposed tensor.

fn int_arange_step( range: Range<i64>, step: usize, device: &<B as Backend>::Device, ) -> <B as Backend>::IntTensorPrimitive

Creates a new tensor with values from the given range with the given step size.

Arguments

range - The range of values.

step - The step size.

device - The device to create the tensor on.

Returns

The tensor with the given values.

fn int_arange( range: Range<i64>, device: &<B as Backend>::Device, ) -> <B as Backend>::IntTensorPrimitive

Creates a new tensor with values from the given range.

Arguments

range - The range of values.

device - The device to create the tensor on.

Returns

The tensor with the given values.

Remarks

Uses arange_step with a step size of 1 under the hood.

fn int_any( tensor: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::BoolTensorPrimitive

Tests if any element in the int tensor evaluates to True.

Arguments

tensor - The tensor to test.

Returns

A boolean tensor with a single element, True if any element in the tensor is True, False otherwise.

fn int_any_dim( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, ) -> <B as Backend>::BoolTensorPrimitive

Tests if any element in the int tensor evaluates to True along a given dimension dim.

Arguments

tensor - The tensor to test.

dim - The axis along which to test.

Returns

A boolean tensor Tensor<B, D, Bool> with the same size as input tensor, except in the dim axis where the size is 1. The elem in the dim axis is True if any element along this dim in the input evaluates to True, False otherwise.

fn int_all( tensor: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::BoolTensorPrimitive

Tests if all elements in the int tensor evaluate to True.

Arguments

tensor - The tensor to test.

Returns

A boolean tensor Tensor<B, 1, Bool> with a single element, True if all elements in the input tensor evaluate to True, False otherwise.

fn int_all_dim( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, ) -> <B as Backend>::BoolTensorPrimitive

Tests if all elements in the int tensor evaluate to True along a given dimension dim.

Arguments

tensor - The tensor to test.

dim - The axis along which to test.

Returns

A boolean tensor Tensor<B, D, Bool> with the same size as input tensor, except in the dim axis where the size is 1. The elem in the dim axis is True if all elements along this dim in the input evaluates to True, False otherwise.

fn int_sign( tensor: <B as Backend>::IntTensorPrimitive, ) -> <B as Backend>::IntTensorPrimitive

Returns the signs of the int tensor.

Arguments

tensor - The tensor to extract the signs from.

Returns

A tensor with the same shape as tensor containing the signs of the elements of tensor.

fn int_sort( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, descending: bool, ) -> <B as Backend>::IntTensorPrimitive

Sort the elements of the input tensor by value along a given dimension.

This sort is unstable (i.e., may reorder equal elements).

Arguments

tensor - The input tensor.

dim - The axis along which to sort.

descending - The sorting order.

Returns

A tensor with the same shape as the input tensor, where the elements are sorted by value.

fn int_sort_with_indices( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, descending: bool, ) -> (<B as Backend>::IntTensorPrimitive, <B as Backend>::IntTensorPrimitive)

Sort the elements of the input tensor by value along a given dimension.

This sort is unstable (i.e., may reorder equal elements).

Arguments

tensor - The input tensor.

dim - The axis along which to sort.

Returns

A tensor with the same shape as the input tensor and corresponding indices, where the elements are sorted by value and the indices map back to the original input tensor.

fn int_argsort( tensor: <B as Backend>::IntTensorPrimitive, dim: usize, descending: bool, ) -> <B as Backend>::IntTensorPrimitive

Returns the indices that sort the elements of the input tensor by value along a given dimension.

This sort is unstable (i.e., may reorder equal elements).

Arguments

tensor - The input tensor.

dim - The axis along which to sort.

descending - The sorting order.

Returns

A tensor with the same shape as the input tensor the indices map back to the original input tensor.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.