burn::tensor

Module repr

Burn Tensor representaton

Structs

AdaptiveAvgPool1dBackwardDescription

AdaptiveAvgPool1dDescription

AdaptiveAvgPool2dBackwardDescription

AdaptiveAvgPool2dDescription

AvgPool1dBackwardDescription

AvgPool1dDescription

AvgPool2dBackwardDescription

AvgPool2dDescription

BinaryOperationDescription

CatOperationDescription

ClampOperationDescription

Conv1dDescription

Conv1dOptionsDescription

Conv2dDescription

Conv2dOptionsDescription

Conv3dDescription

Conv3dOptionsDescription

ConvTranspose1dDescription

ConvTranspose1dOptionsDescription

ConvTranspose2dDescription

ConvTranspose2dOptionsDescription

ConvTranspose3dDescription

ConvTranspose3dOptionsDescription

CustomOpDescription

Custom operation in fusion stream, declaring it’s inputs and outputs.

DeformConv2dBackwardDescription

DeformConv2dDescription

DeformableConv2dOptionsDescription

DequantizeOperationDescription

EmbeddingBackwardDescription

EmbeddingDescription

ExpandDescription

ExpandOperationDescription

Expand operation description.

FlipOperationDescription

Flip operation description.

GatherOperationDescription

HandleContainer

Keep all tensor handles in one place and ensure that all resources are used optimally.

InterpolateBackwardDescription

InterpolateDescription

InterpolateOptionsDescription

MaskFillOperationDescription

MaskWhereOperationDescription

MaxPool1dDescription

MaxPool1dWithIndicesBackwardDescription

MaxPool1dWithIndicesDescription

MaxPool2dDescription

MaxPool2dWithIndicesBackwardDescription

MaxPool2dWithIndicesDescription

PermuteOperationDescription

Permute operation description.

QuantizationParametersDescription

Quantization parameters description.

QuantizeOperationDescription

RandomOperationDescription

ReduceDimWithIndicesDescription

RepeatDimOperationDescription

ReshapeDescription

ScalarOperationDescription

ScatterOperationDescription

SelectAssignOperationDescription

SelectOperationDescription

SliceAssignOperationDescription

SliceOperationDescription

SwapDimsDescription

Swap dim operation description.

TensorDescription

A tensor definition represents a snapshot of a tensor when it was used.

TensorHandle

A tensor representation containing a reference to a tensor resource with a given shape.

TensorId

The tensor unique identifier.

UnaryOperationDescription

Enums

BaseOperationDescription

Basic operations that can be done on any tensor type.

BoolOperationDescription

Operation description specific to a bool tensor.

FloatOperationDescription

Operation description specific to a float tensor.

Handle

Backend tensor handle wrapper tracking their creation state

HandleKind

Handle which points to a backend tensor primitive kind.

IntOperationDescription

Operation description specific to an int tensor.

InterpolateModeDescription

ModuleOperationDescription

Operation description specific to module.

NumericOperationDescription

Numeric operations on int and float tensors.

OperationDescription

Describe all tensor operations possible.

TensorStatus

The status of the current tensor.

Traits

ReprBackend

Backend extension trait that allows an existing backend to use the Burn tensor representation for compilation purpose or other…