burn::tensor

Struct Data

pub struct Data<E, const D: usize> {
    pub value: Vec<E>,
    pub shape: Shape,
}

๐Deprecated since 0.14.0: the internal data format has changed, please use TensorData instead

Data structure for tensors.

Fields
value: Vec<E>
๐Deprecated since 0.14.0: the internal data format has changed, please use TensorData instead

The values of the tensor.

shape: Shape
๐Deprecated since 0.14.0: the internal data format has changed, please use TensorData instead

The shape of the tensor.

Implementations

impl<E, const D: usize> Data<E, D>

pub fn new(value: Vec<E>, shape: Shape) -> Data<E, D>

Constructs a new Data.

impl<const D: usize, E> Data<E, D>
where E: Element,

pub fn convert<EOther>(self) -> Data<EOther, D>
where EOther: Element,

Converts the data to a different element type.

pub fn assert_within_range<EOther>(&self, range: Range<EOther>)
where EOther: Element,

Asserts each value is within a given range.

Arguments

range - The range.

Panics

If any value is not within the half-open range bounded inclusively below and exclusively above (start..end).

impl<E, const D: usize> Data<E, D>
where E: Element,

pub fn random<R>( shape: Shape, distribution: Distribution, rng: &mut R, ) -> Data<E, D>
where R: RngCore,

Populates the data with random values.

impl<E, const D: usize> Data<E, D>
where E: Debug + Element,

pub fn zeros<S>(shape: S) -> Data<E, D>
where S: Into<Shape>,

Populates the data with zeros.

impl<E, const D: usize> Data<E, D>
where E: Debug + Element,

pub fn ones(shape: Shape) -> Data<E, D>

Populates the data with ones.

impl<E, const D: usize> Data<E, D>
where E: Debug + Element,

pub fn full(shape: Shape, fill_value: E) -> Data<E, D>

Populates the data with the given value

impl<E, const D: usize> Data<E, D>
where E: Debug + Copy,

pub fn serialize(&self) -> DataSerialize<E>

Serializes the data.

Returns

The serialized data.

impl<E, const D: usize> Data<E, D>
where E: Into<f64> + Clone + Debug + PartialEq + Element,

pub fn assert_approx_eq(&self, other: &Data<E, D>, precision: usize)

Asserts the data is approximately equal to another data.

Arguments

other - The other data.

precision - The precision of the comparison.

Panics

Panics if the data is not approximately equal.

pub fn assert_approx_eq_diff(&self, other: &Data<E, D>, tolerance: f64)

Asserts the data is approximately equal to another data.

Arguments

other - The other data.

tolerance - The tolerance of the comparison.

Panics

Panics if the data is not approximately equal.

impl<const D: usize> Data<usize, D>

pub fn from_usize<O>(self) -> Data<O, D>
where O: FromPrimitive,

Converts the usize data to a different element type.