mabor

Module train

Train module

Modules

checkpoint

The checkpoint module.

logger

The logger module.

metric

The metric module.

renderer

Renderer modules to display metrics and training information.

train

The trainer module.

Structs

ClassificationOutput

Simple classification output adapted for multiple metrics.

FileApplicationLoggerInstaller

This struct is used to install a local file application logger to output logs to a given file path.

Learner

Learner struct encapsulating all components necessary to train a Neural Network model.

LearnerBuilder

Struct to configure and create a learner.

LearnerSummary

Detailed training summary.

MetricEarlyStoppingStrategy

An early stopping strategy based on a metrics collected during training or validation.

MetricEntry

Contains the metric value at a given time.

MetricSummary

Contains the summary of recorded values for a given metric.

MultiDevicesTrainStep

Multi devices train step.

MultiLabelClassificationOutput

Multi-label classification output adapted for multiple metrics.

RegressionOutput

Simple regression output adapted for multiple metrics.

SummaryMetrics

Contains the summary of recorded metrics for the training and validation steps.

TrainEpoch

A training epoch.

TrainOutput

A training output.

TrainingInterrupter

A handle that allows aborting the training process early.

ValidEpoch

A validation epoch.

Enums

StoppingCondition

The condition that early stopping strategies should follow.

Traits

ApplicationLoggerInstaller

This trait is used to install an application logger.

EarlyStoppingStrategy

A strategy that checks if the training should be stopped.

TrainStep

Trait to be implemented for training models.

ValidStep

Trait to be implemented for validating models.