mabor::train

Module checkpoint

The checkpoint module.

Structs

AsyncCheckpointer

Async checkpointer.

ComposedCheckpointingStrategy

Compose multiple checkpointing strategy and only delete checkpoints when both strategy flag an epoch to be deleted.

ComposedCheckpointingStrategyBuilder

Help building a checkpointing strategy by combining multiple ones.

FileCheckpointer

The file checkpointer.

KeepLastNCheckpoints

Keep the last N checkpoints.

MetricCheckpointingStrategy

Keep the best checkpoint based on a metric.

Enums

CheckpointerError

The error type for checkpointer.

CheckpointingAction

Action to be taken by a checkpointer.

Traits

Checkpointer

The trait for checkpointer.

CheckpointingStrategy

Define when checkpoint should be saved and deleted.