mabor::train

Module metric

The metric module.

Modules

state

State module.

store

Module responsible to save and exposes data collected during training.

Structs

AccuracyInput

The accuracy metric input type.

AccuracyMetric

The accuracy metric.

AurocInput

The AUROC metric input type.

AurocMetric

The Area Under the Receiver Operating Characteristic Curve (AUROC, also referred to as ROC AUC) for binary classification.

ConfusionStatsInput

Input for confusion statistics error types.

CpuMemory

Memory information

CpuTemperature

CPU Temperature in celsius degrees

CpuUse

General CPU Usage metric

CudaMetric

Track basic cuda infos.

FBetaScoreMetric

The F-beta score metric.

HammingScore

The hamming score, sometimes referred to as multi-label or label-based accuracy.

HammingScoreInput

The hamming score input type.

IterationSpeedMetric

The loss metric.

LearningRateMetric

Track the learning rate across iterations.

LossInput

The loss metric input type.

LossMetric

The loss metric.

MetricEntry

Data type that contains the current state of a metric at a given time.

MetricMetadata

Metric metadata that can be used when computing metrics.

PrecisionMetric

The Precision Metric

RecallMetric

The Recall Metric

TopKAccuracyInput

The top-k accuracy metric input type.

TopKAccuracyMetric

The Top-K accuracy metric.

Enums

ClassReduction

The reduction strategy for classification metrics.

NumericEntry

Numeric metric entry.

Traits

Adaptor

Adaptor are used to transform types so that they can be used by metrics.

ItemLazy

Items that are lazy are not ready to be processed by metrics.

Metric

Metric trait.

Numeric

Declare a metric to be numeric.

Functions

format_float

Format a float with the given precision. Will use scientific notation if necessary.