Module burn::train::metric::store

Module responsible to save and exposes data collected during training.

Structs

EventStoreClient

Type that allows to communicate with an event store.

MetricsUpdate

Contains all metric information.

Enums

Aggregate

How to aggregate the metric.

Direction

The direction of the query.

Event

Event happening during the training/validation process.

Split

The split to use.

Traits

EventStore

Defines how training and validation events are collected and searched.