mabor::train::metric

Trait Adaptor

pub trait Adaptor<T> {
    // Required method
    fn adapt(&self) -> T;
}

Adaptor are used to transform types so that they can be used by metrics.

This should be implemented by a model’s output type for all metric inputs that are registered with the learner builder .

Required Methods

fn adapt(&self) -> T

Adapt the type to be passed to a metric.