mabor::train::metric

Trait ItemLazy

pub trait ItemLazy: Send {
    type ItemSync: Send;

    // Required method
    fn sync(self) -> Self::ItemSync;
}

Items that are lazy are not ready to be processed by metrics.

We want to sync them on a different thread to avoid blocking training.

Required Associated Types

type ItemSync: Send

Item that is properly synced and ready to be processed by metrics.

Required Methods

fn sync(self) -> Self::ItemSync

Sync the item.