Crate burn

Burn

This library strives to serve as a comprehensive deep learning framework, offering exceptional flexibility and written in Rust. The main objective is to cater to both researchers and practitioners by simplifying the process of experimenting, training, and deploying models.

Modules

config

The configuration module.

data

Data module.

grad_clipping

Gradient clipping module.

lr_scheduler

Learning rate scheduler module.

module

Module for the neural network module.

nn

Neural network module.

optim

Optimizer module.

record

Module for the recorder.

tensor

Module for the tensor.

train

Train module

Macros

constant

Constant macro.

Type Definitions

LearningRate

Type alias for the learning rate.