Trait burn::backend::autodiff::ops::Backward

pub trait Backward<B, const D: usize, const N: usize>: Send + Sync + Debug + Sized + 'staticwhere
    B: Backend,{
    type State: Clone + Send + Sync + Debug + 'static;

    // Required method
    fn backward(self, ops: Ops<Self::State, N>, grads: &mut Gradients);

    // Provided method
    fn prepare(
        self,
        nodes: [Arc<Node>; N],
        graphs: [Graph; N]
    ) -> OpsPrep<Self, B, Self::State, D, N> { ... }
}

Trait for all operations.

Notes

Concrete types implementing this trait should not have any state. If a state is necessary during the backward pass, they should be declared with the associated type ‘State’.

Required Associated Types

type State: Clone + Send + Sync + Debug + 'static

Associated type to compute the backward pass.

Required Methods

fn backward(self, ops: Ops<Self::State, N>, grads: &mut Gradients)

The backward pass.

Provided Methods

fn prepare( self, nodes: [Arc<Node>; N], graphs: [Graph; N] ) -> OpsPrep<Self, B, Self::State, D, N>

Prepare the backward ops.

Object Safety

This trait is not object safe.