Module burn::module

Module for the neural network module.

Structs

ConstantRecord

Record used for constant type implementing the module trait.

Param

Define a parameter.

ParamId

Parameter ID.

RunningState

A state that can be updated during the forward pass while being thread safe.

Traits

AutodiffModule

Module with auto-differentiation backend.

Module

Trait for all neural network modules.

ModuleMapper

Module mapper trait.

ModuleVisitor

Module visitor trait.

Functions

list_param_ids

List all the parameter ids in a module.

Type Aliases

Devices

Type alias to Vec<B::Device> which supports no_std environments, but automatically using the alloc crate.

Derive Macros

Module

Derive macro for the module.