burn::module

Trait ModuleDisplay

pub trait ModuleDisplay: ModuleDisplayDefault {
    // Provided methods
    fn format(&self, passed_settings: DisplaySettings) -> String { ... }
    fn custom_settings(&self) -> Option<DisplaySettings> { ... }
    fn custom_content(&self, _content: Content) -> Option<Content> { ... }
}

Trait to implement custom display settings for a module.

In order to implement custom display settings for a module,

Add #[module(custom_display)] attribute to the module struct after #[derive(Module)]

Implement ModuleDisplay trait for the module

Provided Methods

fn format(&self, passed_settings: DisplaySettings) -> String

Formats the module with provided display settings.

Arguments

passed_settings - Display settings passed to the module.

Returns

A string representation of the formatted module.

fn custom_settings(&self) -> Option<DisplaySettings>

Custom display settings for the module.

Returns

An optional display settings object.

fn custom_content(&self, _content: Content) -> Option<Content>

Custom attributes for the module.

Arguments

_content - The content object that contains display settings and attributes.

Returns

An optional content object containing the custom attributes.