burn::nn

Module pool

Pooling module

Structs

AdaptiveAvgPool1d

Applies a 1D adaptive avg pooling over input tensors.

AdaptiveAvgPool1dConfig

Configuration to create a 1D adaptive avg pooling layer using the init function.

AdaptiveAvgPool2d

Applies a 2D adaptive avg pooling over input tensors.

AdaptiveAvgPool2dConfig

Configuration to create a 2D adaptive avg pooling layer using the init function.

AvgPool1d

Applies a 1D avg pooling over input tensors.

AvgPool1dConfig

Configuration to create a 1D avg pooling layer using the init function.

AvgPool2d

Applies a 2D avg pooling over input tensors.

AvgPool2dConfig

Configuration to create a 2D avg pooling layer using the init function.

MaxPool1d

Applies a 1D max pooling over input tensors.

MaxPool1dConfig

Configuration to create a 1D max pooling layer using the init function.

MaxPool2d

Applies a 2D max pooling over input tensors.

MaxPool2dConfig

Configuration to create a 2D max pooling layer using the init function.