Trait burn::record::FileRecorder

pub trait FileRecorder: Recorder<RecordArgs = PathBuf, RecordOutput = (), LoadArgs = PathBuf> {
    // Required method
    fn file_extension() -> &'static str;
}

Recorder trait specialized to save and load data to and from files.

Required Methods

fn file_extension() -> &'static str

File extension of the format used by the recorder.