Trait burn::record::FileRecorder

pub trait FileRecorder<B>: Recorder<B, RecordArgs = PathBuf, RecordOutput = (), LoadArgs = PathBuf>where
    B: Backend,{
    // Required method
    fn file_extension() -> &'static str;
}

Recorder trait specialized to save and load data to and from files.

Required Methods

fn file_extension() -> &'static str

File extension of the format used by the recorder.

Object Safety

This trait is not object safe.