Trait burn::record::Record

pub trait Record: Send + Sync {
    type Item<S: PrecisionSettings>: Serialize + DeserializeOwned;

    // Required methods
    fn into_item<S>(self) -> Self::Item<S>
       where S: PrecisionSettings;
    fn from_item<S>(item: Self::Item<S>) -> Self
       where S: PrecisionSettings;
}

Trait to define a family of types which can be recorded using any settings.

Required Associated Types

type Item<S: PrecisionSettings>: Serialize + DeserializeOwned

Type of the item that can be serialized and deserialized.

Required Methods

fn into_item<S>(self) -> Self::Item<S>where S: PrecisionSettings,

Convert the current record into the corresponding item that follows the given settings.

fn from_item<S>(item: Self::Item<S>) -> Selfwhere S: PrecisionSettings,

Convert the given item into a record.