burn::tensor

Module quantization

Tensor quantization module.

Structs

AffineQuantization

Affine quantization scheme.

CalibrationRange

The observed input calibration range.

MinMaxCalibration

Computes the per-tensor quantization range mapping based on the min and max values.

QParams

The quantization tensor data parameters.

QuantizationParametersPrimitive

The quantization parameters primitive.

QuantizedBytes

Quantized data bytes representation.

SymmetricQuantization

Symmetric quantization scheme.

Enums

QuantizationScheme

Quantization scheme.

QuantizationStrategy

Quantization strategy.

QuantizationType

Quantization data type.

QuantizationTypeExpand

Traits

Calibration

Calibration method used to compute the quantization range mapping.

QTensorPrimitive

Quantized tensor primitive.

Quantization

Quantization scheme to convert elements of a higher precision data type E to a lower precision data type Q and vice-versa.

Functions

pack_i8s_to_u32s

Pack signed 8-bit integer values into a sequence of unsigned 32-bit integers.

unpack_u32s_to_i8s

Unpack 32-bit unsigned integer values into a sequence of signed 8-bit integers.

Type Aliases

QuantizationParameters

The tensor quantization parameters.