Trait burn::train::metric::Metric

pub trait Metric: Send + Sync {
    type Input;

    // Required methods
    fn update(
        &mut self,
        item: &Self::Input,
        metadata: &MetricMetadata
    ) -> MetricEntry;
    fn clear(&mut self);
}

Metric trait.

Notes

Implementations should define their own input type only used by the metric. This is important since some conflict may happen when the model output is adapted for each metric’s input type.

Required Associated Types

type Input

The input type of the metric.

Required Methods

fn update( &mut self, item: &Self::Input, metadata: &MetricMetadata ) -> MetricEntry

Update the metric state and returns the current metric entry.

fn clear(&mut self)

Clear the metric state.