Struct burn::train::LearnerBuilder

pub struct LearnerBuilder<B, T, V, M, O, S>where
    T: Send + Sync + 'static,
    V: Send + Sync + 'static,
    B: AutodiffBackend,
    M: AutodiffModule<B>,
    O: Optimizer<M, B>,
    S: LrScheduler<B>,{ /* private fields */ }

Struct to configure and create a learner.

Implementations

impl<B, T, V, M, O, S> LearnerBuilder<B, T, V, M, O, S>where B: AutodiffBackend, T: Send + Sync + 'static, V: Send + Sync + 'static, M: AutodiffModule<B> + Display + 'static, O: Optimizer<M, B>, S: LrScheduler<B>,

pub fn new(directory: &str) -> LearnerBuilder<B, T, V, M, O, S>

Creates a new learner builder.

Arguments

directory - The directory to save the checkpoints.

pub fn metric_loggers<MT, MV>( self, logger_train: MT, logger_valid: MV ) -> LearnerBuilder<B, T, V, M, O, S>where MT: MetricLogger + 'static, MV: MetricLogger + 'static,

Replace the default metric loggers with the provided ones.

Arguments

logger_train - The training logger.

logger_valid - The validation logger.

pub fn with_checkpointing_strategy<CS>(&mut self, strategy: CS)where CS: CheckpointingStrategy + 'static,

Update the checkpointing_strategy.

pub fn renderer<MR>(self, renderer: MR) -> LearnerBuilder<B, T, V, M, O, S>where MR: MetricsRenderer + 'static,

Replace the default CLI renderer with a custom one.

Arguments

renderer - The custom renderer.

pub fn metric_train<Me>(self, metric: Me) -> LearnerBuilder<B, T, V, M, O, S>where Me: Metric + 'static, T: Adaptor<<Me as Metric>::Input>,

Register a training metric.

pub fn metric_valid<Me>(self, metric: Me) -> LearnerBuilder<B, T, V, M, O, S>where Me: Metric + 'static, V: Adaptor<<Me as Metric>::Input>,

Register a validation metric.

pub fn grads_accumulation( self, accumulation: usize ) -> LearnerBuilder<B, T, V, M, O, S>

Enable gradients accumulation.

Notes

When you enable gradients accumulation, the gradients object used by the optimizer will be the sum of all gradients generated by each backward pass. It might be a good idea to reduce the learning to compensate.

The effect is similar to increasing the batch size and the learning rate by the accumulation amount.

pub fn metric_train_numeric<Me>( self, metric: Me ) -> LearnerBuilder<B, T, V, M, O, S>where Me: Metric + Numeric + 'static, T: Adaptor<<Me as Metric>::Input>,

Register a numeric training metric.

pub fn metric_valid_numeric<Me>( self, metric: Me ) -> LearnerBuilder<B, T, V, M, O, S>where Me: Metric + Numeric + 'static, V: Adaptor<<Me as Metric>::Input>,

Register a numeric validation metric.

pub fn num_epochs(self, num_epochs: usize) -> LearnerBuilder<B, T, V, M, O, S>

The number of epochs the training should last.

pub fn devices( self, devices: Vec<<B as Backend>::Device> ) -> LearnerBuilder<B, T, V, M, O, S>

Run the training loop on multiple devices.

pub fn checkpoint(self, checkpoint: usize) -> LearnerBuilder<B, T, V, M, O, S>

The epoch from which the training must resume.

pub fn interrupter(&self) -> TrainingInterrupter

Provides a handle that can be used to interrupt training.

pub fn early_stopping<Strategy>( self, strategy: Strategy ) -> LearnerBuilder<B, T, V, M, O, S>where Strategy: EarlyStoppingStrategy + 'static,

Register an early stopping strategy to stop the training when the conditions are meet.

pub fn log_to_file(self, enabled: bool) -> LearnerBuilder<B, T, V, M, O, S>

By default, Rust logs are captured and written into experiment.log. If disabled, standard Rust log handling will apply.

pub fn with_file_checkpointer<FR>( self, recorder: FR ) -> LearnerBuilder<B, T, V, M, O, S>where FR: FileRecorder<B> + 'static + FileRecorder<<B as AutodiffBackend>::InnerBackend>, <O as Optimizer<M, B>>::Record: 'static, <M as Module<B>>::Record: 'static, <S as LrScheduler<B>>::Record: 'static,

Register a checkpointer that will save the optimizer, the model and the scheduler to different files.

pub fn build( self, model: M, optim: O, lr_scheduler: S ) -> Learner<LearnerComponentsMarker<B, S, M, O, AsyncCheckpointer<<M as Module<B>>::Record, B>, AsyncCheckpointer<<O as Optimizer<M, B>>::Record, B>, AsyncCheckpointer<<S as LrScheduler<B>>::Record, B>, FullEventProcessor<T, V>, Box<dyn CheckpointingStrategy>>>where <M as Module<B>>::Record: 'static, <O as Optimizer<M, B>>::Record: 'static, <S as LrScheduler<B>>::Record: 'static,

Create the learner from a model and an optimizer. The learning rate scheduler can also be a simple learning rate.